+++ server/src/main/java/com/antigravity/race/states/HeatOver.java
@@ -12,7 +12,7 @@ public class HeatOver implements IRaceState {
   }
 
   @Override
-  public void onLap(int lane, float lapTime) {
+  public void onLap(int lane, double lapTime) {
     System.out.println("Race: Ignored onLap - Heat is over");
   }
